STY: move config-import function to base class
